Add support for ParallelCluster 3.13.2 and bug fixes
Don't configure pyxis unless enabled

Update the head node playbook to not configure pyxis unless it is enabled.

Fix on_compute_node_start.sh to not fail if enroot isn't installed.

Resolves #324

Remove custom on_compute_node_start prolog and epilog scripts

/opt/slurm is not mounted at this point so these custom scripts would not be available to run.
Instead, use the custom on_compute_node_configured prolog and epilog scripts.

Resolves #329

Fix permissions on /var/run

The fix for subuid/subgid changed permissions of /var/run to 1777 instead of 1755.

Resolves #330

Add support for ParallelCluster 3.13.2

Add a dependency on CreateBuildFiles for UpdateHeadNode so that build files get created before updating the head node.

Resolves #331

+++ source/resources/parallel-cluster/config/bin/on_compute_node_start.sh
@@ -30,10 +30,6 @@ trap on_exit EXIT
 
 # /opt/slurm isn't mounted yet.
 
-if [ -e $config_bin_dir/on_compute_node_start_custom_prolog.sh ]; then
-    $config_bin_dir/on_compute_node_start_custom_prolog.sh
-fi
-
 # Configure pyxis and enroot
 
 # Configure Enroot
@@ -44,18 +40,16 @@ sudo mkdir -p $ENROOT_PERSISTENT_DIR
 sudo chmod 1777 $ENROOT_PERSISTENT_DIR
 sudo mkdir -p $ENROOT_VOLATILE_DIR
 sudo chmod 1777 $ENROOT_VOLATILE_DIR
-sudo cp /opt/parallelcluster/examples/enroot/enroot.conf /etc/enroot/enroot.conf
-sudo chmod 0644 /etc/enroot/enroot.conf
+if [[ -e /opt/parallelcluster/examples/enroot/enroot.conf ]] && [[ -d /etc/enroot/ ]]; then
+    sudo cp /opt/parallelcluster/examples/enroot/enroot.conf /etc/enroot/enroot.conf
+    sudo chmod 0644 /etc/enroot/enroot.conf
+fi
 
 # Configure Pyxis
 PYXIS_RUNTIME_DIR="/run/pyxis"
 
 sudo mkdir -p $PYXIS_RUNTIME_DIR
 sudo chmod 1777 $PYXIS_RUNTIME_DIR
-
-if [ -e $config_bin_dir/on_compute_node_start_custom_epilog.sh ]; then
-    $config_bin_dir/on_compute_node_start_custom_epilog.sh
-fi
 
 echo "$(date): Finished ${script_name}"
